fix(TDI-44213) : without migration
