--- workday/src/main/java/org/talend/components/workday/dataset/Migration.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (C) 2006-2020 Talend Inc. - www.talend.com
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package org.talend.components.workday.dataset;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.talend.sdk.component.api.component.MigrationHandler;
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class Migration implements MigrationHandler {
-
-    @Override
-    public Map<String, String> migrate(int incomingVersion, Map<String, String> incomingData) {
-        log.info("Starting Workday Dataset Migration from " + incomingVersion);
-
-        if (incomingVersion < 2) { // 1 or -1
-            final Map<String, String> newConfig = new HashMap<>(incomingData);
-
-            if (!incomingData.containsKey("mode")) {
-                newConfig.put("mode", "WQL"); // if old default value is absent.
-            }
-            return newConfig;
-        }
-        return incomingData;
-    }
-
-}
