+++ workday/src/main/java/org/talend/components/workday/datastore/WorkdayDataStore.java
@@ -13,34 +13,39 @@
 package org.talend.components.workday.datastore;
 
 import java.io.Serializable;
+import java.nio.charset.Charset;
+import java.util.Base64;
 
 import org.talend.components.workday.service.UIActionService;
 import org.talend.sdk.component.api.component.Version;
 import org.talend.sdk.component.api.configuration.Option;
 import org.talend.sdk.component.api.configuration.action.Checkable;
+import org.talend.sdk.component.api.configuration.action.Validable;
 import org.talend.sdk.component.api.configuration.condition.ActiveIf;
 import org.talend.sdk.component.api.configuration.constraint.Required;
 import org.talend.sdk.component.api.configuration.type.DataStore;
 import org.talend.sdk.component.api.configuration.ui.DefaultValue;
 import org.talend.sdk.component.api.configuration.ui.layout.GridLayout;
+import org.talend.sdk.component.api.configuration.ui.widget.Credential;
 import org.talend.sdk.component.api.meta.Documentation;
 
 import lombok.Data;
 
 @Data
-@Version(value = 2, migrationHandler = Migration.class)
+@Version(value = 1)
 @DataStore("WorkdayDataStore")
 @GridLayout({ //
         @GridLayout.Row({ "authentication" }), //
-        @GridLayout.Row({ "clientIdForm", "loginForm" }) //
+        @GridLayout.Row({ "clientId", "clientSecret", "loginForm" }), //
+        @GridLayout.Row({ "tenantAlias" }) //
 })
 @GridLayout(names = GridLayout.FormType.ADVANCED, //
-        value = { @GridLayout.Row({ "clientIdForm", "loginForm" }) })
+        value = { @GridLayout.Row({ "authEndpoint", "loginForm" }), @GridLayout.Row("endpoint") })
 @Checkable(UIActionService.HEALTH_CHECK)
 @Documentation("DataStore for workday connector")
 public class WorkdayDataStore implements Serializable {
 
-    private static final long serialVersionUID = -8628647674176772061L;
+    private static final long serialVersionUID = 8254399046469388027L;
 
     public enum AuthenticationType {
         CLIENT_ID,
@@ -54,18 +59,45 @@ public class WorkdayDataStore implements Serializable {
     private AuthenticationType authentication = AuthenticationType.CLIENT_ID;
 
     @Option
+    @Validable(UIActionService.VALIDATION_URL_PROPERTY)
+    @Documentation("Workday token Auth Endpoint (host only, ie: https://auth.api.workday.com/v1/token)")
+    private String authEndpoint = "https://auth.api.workday.com";
+
+    @Option
     @Documentation("Workday Client Id")
     @ActiveIf(target = "authentication", value = "CLIENT_ID")
-    private ClientIdForm clientIdForm = new ClientIdForm();
+    private String clientId;
+
+    @Option
+    @Credential
+    @Documentation("Workday Client Secret")
+    @ActiveIf(target = "authentication", value = "CLIENT_ID")
+    private String clientSecret;
+
+    @Option
+    @Documentation("Workday endpoint for REST services")
+    @ActiveIf(target = "authentication", value = "CLIENT_ID")
+    private String endpoint = "https://api.workday.com";
+
+    @Option
+    @ActiveIf(target = "authentication", value = "CLIENT_ID")
+    @Documentation("Workday tenant alias")
+    private String tenantAlias;
 
     @Option
     @Documentation("Workday Login connection")
     @ActiveIf(target = "authentication", value = "LOGIN")
     private UserFormForReport loginForm = new UserFormForReport();
 
-    public String getEndPoint() {
+    public String getAuthorizationHeader() {
+        final String idSecret = this.clientId + ':' + this.clientSecret;
+        final String idForHeader = Base64.getEncoder().encodeToString(idSecret.getBytes(Charset.defaultCharset()));
+        return "Basic " + idForHeader;
+    }
+
+    public String getWorkdayEndPoint() {
         if (this.authentication == AuthenticationType.CLIENT_ID) {
-            return this.getClientIdForm().getEndpoint();
+            return this.endpoint;
         } else {
             return this.getLoginForm().getRealEndpoint();
         }
