+++ workday/src/main/resources/org/talend/components/workday/datastore/Messages.properties
@@ -1,24 +1,24 @@
-ClientIdForm.authEndpoint._displayName=Authentication End Point
-ClientIdForm.authEndpoint._placeholder=https://auth.api.workday.com
+WorkdayDataStore.authEndpoint._displayName=Authentication End Point
+WorkdayDataStore.authEndpoint._placeholder=https://auth.api.workday.com
 
-ClientIdForm.clientId._displayName=Client Identifier
-ClientIdForm.clientId._placeholder=
+WorkdayDataStore.clientId._displayName=Client Identifier
+WorkdayDataStore.clientId._placeholder=
 
-ClientIdForm.clientSecret._displayName=Client Secret
-ClientIdForm.clientSecret._placeholder=
+WorkdayDataStore.clientSecret._displayName=Client Secret
+WorkdayDataStore.clientSecret._placeholder=
 
-ClientIdForm.endpoint._displayName=Service Endpoint
-ClientIdForm.endpoint._placeholder=https://api.workday.com
+WorkdayDataStore.endpoint._displayName=Service Endpoint
+WorkdayDataStore.endpoint._placeholder=https://api.workday.com
 
-ClientIdForm.tenantAlias._displayName=Tenant Alias
-ClientIdForm.tenantAlias._placeholder=
+WorkdayDataStore.tenantAlias._displayName=Tenant Alias
+WorkdayDataStore.tenantAlias._placeholder=
 
 WorkdayDataStore.authentication._displayName=Authentication Mode
 AuthenticationType.CLIENT_ID._displayName=Client Id Authentication Mode
 AuthenticationType.LOGIN._displayName   =Login Authentication Mode
 
 Workday.datastore.WorkdayDataStore._displayName=Workday Connection
-WorkdayDataStore.clientIdForm._displayName=ClientId Authentication Form
+
 WorkdayDataStore.loginForm._displayName=Login Authentication Form
 
 UserFormForReport.endpointPattern._displayName=Service Endpoint
