+++ workday/src/test/java/org/talend/components/workday/WorkdayBaseTest.java
@@ -17,7 +17,6 @@ import java.io.InputStream;
 import java.io.UncheckedIOException;
 import java.util.Properties;
 
-import org.talend.components.workday.datastore.ClientIdForm;
 import org.talend.components.workday.datastore.WorkdayDataStore;
 import org.talend.sdk.component.api.DecryptedServer;
 import org.talend.sdk.component.junit5.WithMavenServers;
@@ -52,15 +51,15 @@ public class WorkdayBaseTest {
     protected WorkdayDataStore buildDataStore() {
         Properties props = this.workdayProps();
         WorkdayDataStore wds = new WorkdayDataStore();
-        final ClientIdForm form = wds.getClientIdForm();
-        form.setClientId(serverWorkday.getUsername());
-        form.setClientSecret(serverWorkday.getPassword());
+
+        wds.setClientId(serverWorkday.getUsername());
+        wds.setClientSecret(serverWorkday.getPassword());
 
         // tenant
-        form.setTenantAlias(serverWorkdayTenant.getUsername());
+        wds.setTenantAlias(serverWorkdayTenant.getUsername());
 
-        form.setAuthEndpoint(props.getProperty("authendpoint"));
-        form.setEndpoint(props.getProperty("endpoint"));
+        wds.setAuthEndpoint(props.getProperty("authendpoint"));
+        wds.setEndpoint(props.getProperty("endpoint"));
         return wds;
     }
 }
