+++ workday/src/test/java/org/talend/components/workday/service/AccessTokenProviderTest.java
@@ -57,7 +57,7 @@ class AccessTokenProviderTest extends WorkdayBaseTest {
         AccessTokenProvider provider = factory.create(AccessTokenProvider.class, WorkdayBaseTest.defaultAuthenticationURL);
 
         WorkdayDataStore wds = this.buildDataStore();
-        wds.getClientIdForm().setClientSecret("fautSecret");
+        wds.setClientSecret("fautSecret");
 
         Assertions.assertThrows(WorkdayException.class, () -> provider.getAccessToken(wds));
     }
